fix(web): keep frontend source budgets green
Split API normalization and stylesheet modules after the clay redesign.
