+++ web/src/components/AdminRecentRequestsPanel.tsx
@@ -1,6 +1,5 @@
 import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
-
 import type { QueryLoadState } from '../admin/queryLoadState'
 import type { LogFacetOption, RequestLog, RequestLogBodies } from '../api'
 import type { AdminTranslations } from '../i18n'
@@ -58,16 +57,13 @@ import {
 import SegmentedTabs from './ui/SegmentedTabs'
 import { TableBody, TableCell, TableHead, TableHeader, TableRow } from './ui/table'
 import { Tooltip, TooltipContent, TooltipTrigger } from './ui/tooltip'
-
 type Language = 'en' | 'zh'
 type RecentRequestsVariant = 'admin' | 'token'
 type RecentRequestsOutcomeFilterKind = 'result' | 'keyEffect' | 'bindingEffect' | 'selectionEffect'
-
 export interface RecentRequestsOutcomeFilter {
   kind: RecentRequestsOutcomeFilterKind
   value: string
 }
-
 export interface AdminRecentRequestsPanelProps {
   variant: RecentRequestsVariant
   language: Language
@@ -116,27 +112,22 @@ export interface AdminRecentRequestsPanelProps {
   onOpenToken?: (id: string) => void
   loadLogBodies: (log: RequestLog, signal: AbortSignal) => Promise<RequestLogBodies>
 }
-
 type LogBodiesLoadState =
   | { status: 'loading' }
   | { status: 'ready'; value: RequestLogBodies }
   | { status: 'error'; message: string }
-
 const requestKindBillingQuickFilterOptions = [
   { value: 'all', label: 'Any' },
   { value: 'billable', label: 'Paid' },
   { value: 'non_billable', label: 'Free' },
 ] as const
-
 const requestKindProtocolQuickFilterOptions = [
   { value: 'all', label: 'Any' },
   { value: 'mcp', label: 'MCP' },
   { value: 'api', label: 'API' },
 ] as const
-
 const recentRequestsAllFilterValue = '__all__'
 const recentRequestsCompactAllLabel = 'All'
-
 function statusTone(status: string): StatusTone {
   const normalized = status.trim().toLowerCase()
   if (normalized === 'success') return 'success'
@@ -144,7 +135,6 @@ function statusTone(status: string): StatusTone {
   if (normalized === 'error') return 'error'
   return 'neutral'
 }
-
 function statusLabel(status: string, strings: AdminTranslations): string {
   const normalized = status.trim().toLowerCase()
   if (normalized === 'success') return strings.statuses.success
@@ -153,7 +143,6 @@ function statusLabel(status: string, strings: AdminTranslations): string {
   if (normalized === 'quota_exhausted') return strings.statuses.quota_exhausted
   return status || strings.logs.errors.none
 }
-
 function keyEffectTone(code: string | null | undefined): StatusTone {
   switch ((code ?? '').trim()) {
     case 'quarantined':
@@ -175,7 +164,6 @@ function keyEffectTone(code: string | null | undefined): StatusTone {
       return 'neutral'
   }
 }
-
 function keyEffectLabel(code: string | null | undefined, strings: AdminTranslations): string {
   switch ((code ?? '').trim()) {
     case 'quarantined':
@@ -203,11 +191,9 @@ function keyEffectLabel(code: string | null | undefined, strings: AdminTranslati
       return strings.logs.keyEffects.unknown
   }
 }
-
 function keyEffectBadgeLabel(log: RequestLog, strings: AdminTranslations): string {
   return keyEffectLabel(log.key_effect_code, strings)
 }
-
 function keyEffectBadgeCompactLabel(log: RequestLog, strings: AdminTranslations, language: Language): string {
   switch ((log.key_effect_code ?? '').trim()) {
     case 'quarantined':
@@ -228,7 +214,6 @@ function keyEffectBadgeCompactLabel(log: RequestLog, strings: AdminTranslations,
       return keyEffectBadgeLabel(log, strings)
   }
 }
-
 function bindingEffectTone(code: string | null | undefined): StatusTone {
   const apiRebalanceTone = apiRebalanceBindingEffectTone(code)
   if (apiRebalanceTone) return apiRebalanceTone
@@ -243,7 +228,6 @@ function bindingEffectTone(code: string | null | undefined): StatusTone {
       return 'neutral'
   }
 }
-
 function bindingEffectLabel(code: string | null | undefined, strings: AdminTranslations): string {
   const apiRebalanceLabel = apiRebalanceBindingEffectLabel(code, strings)
   if (apiRebalanceLabel) return apiRebalanceLabel
@@ -261,7 +245,6 @@ function bindingEffectLabel(code: string | null | undefined, strings: AdminTrans
       return strings.logs.bindingEffects.unknown
   }
 }
-
 function bindingEffectBadgeLabel(log: RequestLog, strings: AdminTranslations, language: Language): string {
   const apiRebalanceLabel = apiRebalanceBindingEffectBadgeLabel(log.binding_effect_code, language)
   if (apiRebalanceLabel) return apiRebalanceLabel
@@ -276,7 +259,6 @@ function bindingEffectBadgeLabel(log: RequestLog, strings: AdminTranslations, la
       return bindingEffectLabel(log.binding_effect_code, strings)
   }
 }
-
 function selectionEffectTone(code: string | null | undefined): StatusTone {
   const apiRebalanceTone = apiRebalanceSelectionEffectTone(code)
   if (apiRebalanceTone) return apiRebalanceTone
@@ -294,7 +276,6 @@ function selectionEffectTone(code: string | null | undefined): StatusTone {
       return 'neutral'
   }
 }
-
 function selectionEffectLabel(code: string | null | undefined, strings: AdminTranslations): string {
   const apiRebalanceLabel = apiRebalanceSelectionEffectLabel(code, strings)
   if (apiRebalanceLabel) return apiRebalanceLabel
@@ -318,7 +299,6 @@ function selectionEffectLabel(code: string | null | undefined, strings: AdminTra
       return strings.logs.selectionEffects.unknown
   }
 }
-
 function selectionEffectBadgeLabel(log: RequestLog, strings: AdminTranslations, language: Language): string {
   const apiRebalanceLabel = apiRebalanceSelectionEffectBadgeLabel(log.selection_effect_code, language)
   if (apiRebalanceLabel) return apiRebalanceLabel
@@ -339,7 +319,6 @@ function selectionEffectBadgeLabel(log: RequestLog, strings: AdminTranslations, 
       return selectionEffectLabel(log.selection_effect_code, strings)
   }
 }
-
 function formatKeyEffectSummary(
   log: RequestLog,
   strings: AdminTranslations,
@@ -375,7 +354,6 @@ function formatKeyEffectSummary(
       return summary && summary.length > 0 ? summary : strings.logDetails.noKeyEffect
   }
 }
-
 function formatBindingEffectSummary(
   log: RequestLog,
   strings: AdminTranslations,
@@ -403,7 +381,6 @@ function formatBindingEffectSummary(
       return summary && summary.length > 0 ? summary : strings.logDetails.noBindingEffect
   }
 }
-
 function formatSelectionEffectSummary(
   log: RequestLog,
   strings: AdminTranslations,
@@ -443,12 +420,10 @@ function formatSelectionEffectSummary(
       return summary && summary.length > 0 ? summary : strings.logDetails.noSelectionEffect
   }
 }
-
 function hasExplicitEffect(code: string | null | undefined): boolean {
   const normalized = (code ?? '').trim()
   return normalized.length > 0 && normalized !== 'none'
 }
-
 function renderEffectBadges(
   log: RequestLog,
   strings: AdminTranslations,
@@ -457,7 +432,6 @@ function renderEffectBadges(
 ): JSX.Element {
   const badgeClassName = className ? `${className} recent-requests-effect-badge` : 'recent-requests-effect-badge'
   const badges: JSX.Element[] = []
-
   if (hasExplicitEffect(log.key_effect_code)) {
     badges.push(
       <StatusBadge
@@ -494,7 +468,6 @@ function renderEffectBadges(
       </StatusBadge>,
     )
   }
-
   if (badges.length === 0) {
     badges.push(
       <StatusBadge
@@ -507,33 +480,25 @@ function renderEffectBadges(
       </StatusBadge>,
     )
   }
-
   return <div className="recent-requests-effect-badges">{badges}</div>
 }
-
 function formatRequestStatusPair(httpStatus: number | null, mcpStatus: number | null): string {
   return `${httpStatus ?? '—'} / ${mcpStatus ?? '—'}`
 }
-
 function requestStatusPairClassName(log: RequestLog): string { return `status-pair-trigger status-pair-trigger--${statusTone(log.result_status)}` }
-
 function formatRequestStatusTooltip(log: RequestLog, strings: AdminTranslations): string {
   return `${strings.logs.table.httpStatus}: ${log.http_status ?? '—'} · ${strings.logs.table.mcpStatus}: ${log.mcp_status ?? '—'}`
 }
-
 function formatChargedCredits(value: number | null | undefined): string {
   return value != null ? String(value) : '—'
 }
-
 function formatRequestLine(log: RequestLog): string {
   const query = log.query ? `?${log.query}` : ''
   return `${log.method} ${log.path}${query}`
 }
-
 function formatErrorMessage(log: RequestLog, strings: AdminTranslations['logs']['errors']): string {
   const message = log.error_message?.trim()
   if (message) return message
-
   const status = log.result_status.toLowerCase()
   if (status === 'quota_exhausted') {
     if (log.http_status != null) {
@@ -541,7 +506,6 @@ function formatErrorMessage(log: RequestLog, strings: AdminTranslations['logs'][
     }
     return strings.quotaExhausted
   }
-
   if (status === 'error') {
     if (log.http_status != null && log.mcp_status != null) {
       return strings.requestFailedHttpMcp
@@ -552,12 +516,10 @@ function formatErrorMessage(log: RequestLog, strings: AdminTranslations['logs'][
     if (log.mcp_status != null) return strings.requestFailedMcp.replace('{mcp}', String(log.mcp_status))
     return strings.requestFailedGeneric
   }
-
   if (status === 'success') return strings.none
   if (log.http_status != null) return strings.httpStatus.replace('{http}', String(log.http_status))
   return strings.none
 }
-
 function summarizeSingleFacet(
   selectedValue: string | null | undefined,
   options: LogFacetOption[] | undefined,
@@ -567,7 +529,6 @@ function summarizeSingleFacet(
   if (!normalized) return fallbackLabel
   return options?.find((option) => option.value === normalized)?.value ?? normalized
 }
-
 function summarizeOutcomeFilter(
   filter: RecentRequestsOutcomeFilter | null,
   strings: AdminTranslations,
@@ -585,7 +546,6 @@ function summarizeOutcomeFilter(
   }
   return selectionEffectLabel(filter.value, strings)
 }
-
 function renderOutcomeFacetLabel(
   kind: RecentRequestsOutcomeFilterKind,
   value: string,
@@ -609,7 +569,6 @@ function renderOutcomeFacetLabel(
           : selectionEffectLabel(value, strings)
   return <StatusBadge tone={tone}>{label}</StatusBadge>
 }
-
 function summarizeRequestKindTrigger(
   effectiveSelectedRequestKinds: string[],
   hasActiveQuickRequestKindFilters: boolean,
@@ -623,7 +582,6 @@ function summarizeRequestKindTrigger(
   if (effectiveSelectedRequestKinds.length <= 2) return requestKindSummary
   return language === 'zh' ? `已选 ${effectiveSelectedRequestKinds.length} 项` : `${effectiveSelectedRequestKinds.length} selected`
 }
-
 function RecentRequestDetails({
   log,
   logBodiesState,
@@ -698,7 +656,6 @@ function RecentRequestDetails({
       ? { label: strings.logDetails.fallbackReason, value: log.fallback_reason }
       : null,
   ].filter((entry): entry is { label: string; value: string } => entry != null)
-
   return (
     <div className="log-details-panel">
       <div className="log-details-summary">
@@ -806,7 +763,6 @@ function RecentRequestDetails({
     </div>
   )
 }
-
 export default function AdminRecentRequestsPanel({
   variant,
   language,
@@ -856,7 +812,6 @@ export default function AdminRecentRequestsPanel({
   const [logBodiesById, setLogBodiesById] = useState<Record<number, LogBodiesLoadState>>({})
   const [headerFiltersTarget, setHeaderFiltersTarget] = useState<HTMLElement | null>(null)
   const logBodyControllersRef = useRef<Map<number, AbortController>>(new Map())
-
   useEffect(() => {
     if (!headerFiltersTargetId) {
       setHeaderFiltersTarget(null)
@@ -864,7 +819,6 @@ export default function AdminRecentRequestsPanel({
     }
     setHeaderFiltersTarget(document.getElementById(headerFiltersTargetId))
   }, [headerFiltersTargetId])
-
   useEffect(() => {
     setExpandedLogs(new Set())
     for (const controller of logBodyControllersRef.current.values()) {
@@ -873,19 +827,16 @@ export default function AdminRecentRequestsPanel({
     logBodyControllersRef.current.clear()
     setLogBodiesById({})
   }, [loadLogBodies, logs])
-
   const triggerLoadLogBodies = useCallback(
     (log: RequestLog, force = false) => {
       const currentState = logBodiesById[log.id]
       if (!force && (currentState?.status === 'loading' || currentState?.status === 'ready')) {
         return
       }
-
       logBodyControllersRef.current.get(log.id)?.abort()
       const controller = new AbortController()
       logBodyControllersRef.current.set(log.id, controller)
       setLogBodiesById((current) => ({ ...current, [log.id]: { status: 'loading' } }))
-
       loadLogBodies(log, controller.signal)
         .then((value) => {
           if (controller.signal.aborted) return
@@ -915,7 +866,6 @@ export default function AdminRecentRequestsPanel({
     },
     [loadLogBodies, logBodiesById, strings.logDetails.loadBodyFailed],
   )
-
   const toggleExpandedLog = useCallback(
     (log: RequestLog) => {
       const expanded = expandedLogs.has(log.id)
@@ -934,14 +884,12 @@ export default function AdminRecentRequestsPanel({
     },
     [expandedLogs, triggerLoadLogBodies],
   )
-
   const retryLoadBodies = useCallback(
     (log: RequestLog) => {
       triggerLoadLogBodies(log, true)
     },
     [triggerLoadLogBodies],
   )
-
   const normalizedSelectedRequestKinds = useMemo(
     () => Array.from(new Set(selectedRequestKinds.map((value) => value.trim()).filter(Boolean))),
     [selectedRequestKinds],
@@ -1112,7 +1060,6 @@ export default function AdminRecentRequestsPanel({
           </DropdownMenuContent>
         </DropdownMenu>
       </div>
-
       <div className="recent-requests-filter-field">
         <span className="recent-requests-filter-label">{strings.logs.filters.resultOrEffect}</span>
         <Select
@@ -1234,7 +1181,6 @@ export default function AdminRecentRequestsPanel({
           </SelectContent>
         </Select>
       </div>
-
       {showKeyColumn && onKeyFilterChange ? (
         <div className="recent-requests-filter-field">
           <span className="recent-requests-filter-label">{strings.logs.table.key}</span>
@@ -1260,7 +1206,6 @@ export default function AdminRecentRequestsPanel({
   const headerFiltersPortal = headerFiltersTarget
     ? createPortal(renderFilters('recent-requests-filters--header'), headerFiltersTarget)
     : null
-
   return (
     <section className="surface panel">
       {headerFiltersPortal}
@@ -1273,7 +1218,6 @@ export default function AdminRecentRequestsPanel({
         ) : null}
         {renderFilters(headerFiltersTarget ? 'recent-requests-filters--mobile-only' : undefined)}
       </div>
-
       <AdminTableShell
         className={desktopClassName}
         tableClassName={`recent-requests-table recent-requests-table--${variant}`}
@@ -1455,7 +1399,6 @@ export default function AdminRecentRequestsPanel({
           )}
         </TableBody>
       </AdminTableShell>
-
       <AdminLoadingRegion
         className={mobileClassName}
         loadState={loadState}
@@ -1603,7 +1546,6 @@ export default function AdminRecentRequestsPanel({
           })
         )}
       </AdminLoadingRegion>
-
       <AdminTablePagination
         page={1}
         totalPages={1}
