+++ web/src/i18n/translations/en.ts
@@ -569,21 +569,12 @@ export const EN: TranslationShape = {
       },
       systemSettings: {
         title: 'System Settings', description: 'Adjust the global request-rate threshold, MCP session affinity, Rebalance behavior, and blocked-key base limit without restarting.',
-        helpLabel: 'Show system settings help',
-        form: {
-          title: 'MCP gateway controls',
-          description: 'Manage the global request-rate threshold, the MCP session affinity pool, the Rebalance MCP rollout ratio, and the blocked-key base limit in one place.',
-          displayDensityTitle: 'List density',
-          displayDensityDescription: 'Unifies admin list row height, type size, and cell spacing.',
-          displayDensityComfortable: 'Comfortable',
-          displayDensityCompact: 'Compact',
-          displayDensityStoredHint: 'Saved only in this browser and applied immediately.',
-          accessDisplayTitle: 'Access and display',
-          limitsTitle: 'Global limits and warnings',
-          gatewayTitle: 'MCP gateway controls',
-          gatewaySectionTitle: 'Session affinity and Rebalance',
-          apiRebalanceTitle: 'Tavily API Rebalance',
-          requestRateLimitLabel: 'Max requests per 5 minutes',
+        helpLabel: 'Show system settings help', form: {
+          title: 'MCP gateway controls', description: 'Manage the global request-rate threshold, the MCP session affinity pool, the Rebalance MCP rollout ratio, and the blocked-key base limit in one place.',
+          displayDensityTitle: 'List density', displayDensityDescription: 'Unifies admin list row height, type size, and cell spacing.',
+          displayDensityComfortable: 'Comfortable', displayDensityCompact: 'Compact', displayDensityStoredHint: 'Saved only in this browser and applied immediately.',
+          accessDisplayTitle: 'Access and display', limitsTitle: 'Global limits and warnings', gatewayTitle: 'MCP gateway controls',
+          gatewaySectionTitle: 'Session affinity and Rebalance', apiRebalanceTitle: 'Tavily API Rebalance', requestRateLimitLabel: 'Max requests per 5 minutes',
           requestRateLimitHint: 'Applies to the global request-rate limiter. Saving takes effect on the next authenticated request, but keeps the current rolling-window counters intact.',
           currentRequestRateLimitValue: 'Current limit: {count}',
           countLabel: 'Affinity key count',
