+++ web/src/i18n/translations/zh.ts
@@ -569,21 +569,12 @@ export const ZH: TranslationShape = {
       systemSettings: {
         title: '系统设置',
         description: '调整系统级 request-rate、MCP session 亲和、Rebalance 策略与封禁数基础值，并且无需重启即可生效。',
-        helpLabel: '显示系统设置说明',
-        form: {
-          title: 'MCP 网关控制',
-          description: '这里统一管理全局 request-rate 阈值、MCP session 亲和池、Rebalance MCP 灰度比例与封禁数基础值。',
-          displayDensityTitle: '列表密度',
-          displayDensityDescription: '统一管理端列表的行高、字号和单元格间距。',
-          displayDensityComfortable: '一般',
-          displayDensityCompact: '紧凑',
-          displayDensityStoredHint: '仅保存在当前浏览器，切换后立即应用。',
-          accessDisplayTitle: '访问与显示',
-          limitsTitle: '全局限制与提示',
-          gatewayTitle: 'MCP 网关控制',
-          gatewaySectionTitle: 'Session 亲和与 Rebalance',
-          apiRebalanceTitle: 'Tavily API Rebalance',
-          requestRateLimitLabel: '5 分钟最大请求数',
+        helpLabel: '显示系统设置说明', form: {
+          title: 'MCP 网关控制', description: '这里统一管理全局 request-rate 阈值、MCP session 亲和池、Rebalance MCP 灰度比例与封禁数基础值。',
+          displayDensityTitle: '列表密度', displayDensityDescription: '统一管理端列表的行高、字号和单元格间距。',
+          displayDensityComfortable: '一般', displayDensityCompact: '紧凑', displayDensityStoredHint: '仅保存在当前浏览器，切换后立即应用。',
+          accessDisplayTitle: '访问与显示', limitsTitle: '全局限制与提示', gatewayTitle: 'MCP 网关控制',
+          gatewaySectionTitle: 'Session 亲和与 Rebalance', apiRebalanceTitle: 'Tavily API Rebalance', requestRateLimitLabel: '5 分钟最大请求数',
           requestRateLimitHint: '作用于全局 request-rate limiter；保存后对下一次已鉴权请求立即生效，但不会清空当前 5 分钟窗口中的已有计数。',
           currentRequestRateLimitValue: '当前阈值：{count}',
           countLabel: '亲和 key 数量',
