+++ web/src/styles/admin.css
@@ -556,224 +556,6 @@
   }
 }
 
-.user-console-tokens-table,
-.user-console-logs-table {
-  width: 100%;
-  min-width: 0;
-  table-layout: fixed;
-}
-
-.user-console-tokens-table th,
-.user-console-tokens-table td,
-.user-console-logs-table th,
-.user-console-logs-table td {
-  vertical-align: top;
-}
-
-.user-console-tokens-table th:nth-child(1),
-.user-console-tokens-table td:nth-child(1) {
-  width: 14%;
-}
-
-.user-console-tokens-table th:nth-child(2),
-.user-console-tokens-table td:nth-child(2) {
-  width: 44%;
-}
-
-.user-console-tokens-table th:nth-child(3),
-.user-console-tokens-table td:nth-child(3) {
-  width: 24%;
-}
-
-.user-console-tokens-table th:nth-child(4),
-.user-console-tokens-table td:nth-child(4) {
-  width: 18%;
-}
-
-.user-console-logs-table th:nth-child(1),
-.user-console-logs-table td:nth-child(1) {
-  width: 44%;
-  min-width: 0;
-}
-
-.user-console-logs-table th:nth-child(2),
-.user-console-logs-table td:nth-child(2) {
-  width: 16%;
-  min-width: 0;
-  text-align: left;
-}
-
-.user-console-logs-table th:nth-child(3),
-.user-console-logs-table td:nth-child(3) {
-  width: 40%;
-  min-width: 0;
-  text-align: left;
-}
-
-.user-console-logs-table td {
-  white-space: normal;
-  padding-top: 8px;
-  padding-bottom: 8px;
-}
-
-.user-console-cell-stack {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.user-console-cell-item {
-  display: flex;
-  align-items: baseline;
-  justify-content: space-between;
-  gap: 12px;
-  padding-bottom: 4px;
-  border-bottom: 1px dashed hsl(var(--border) / 0.55);
-  font-size: 0.9rem;
-}
-
-.user-console-cell-item:last-child {
-  border-bottom: none;
-  padding-bottom: 0;
-}
-
-.user-console-cell-item span {
-  color: hsl(var(--muted-foreground));
-}
-
-.user-console-cell-item strong {
-  font-family: 'JetBrains Mono', 'SFMono-Regular', Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace;
-  font-weight: 700;
-  color: hsl(var(--foreground));
-}
-
-.user-console-tokens-table .table-actions {
-  justify-content: flex-end;
-  align-items: flex-start;
-  gap: 8px;
-  flex-wrap: wrap;
-}
-
-.user-console-tokens-table .table-actions button {
-  white-space: nowrap;
-}
-
-.user-console-log-stack {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 2px;
-  min-width: 0;
-}
-
-.user-console-log-main {
-  font-size: 0.82rem;
-  font-weight: 700;
-  color: hsl(var(--foreground));
-}
-
-.user-console-push-status-slot {
-  display: flex;
-  min-width: 1.75rem;
-  min-height: 1.75rem;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.user-console-push-status-trigger {
-  display: inline-flex;
-  width: 1.75rem;
-  height: 1.75rem;
-  align-items: center;
-  justify-content: center;
-  border: 0;
-  border-radius: 999px;
-  background: transparent;
-  color: hsl(var(--warning));
-  cursor: help;
-  transition:
-    color 160ms ease,
-    background-color 160ms ease,
-    opacity 160ms ease;
-}
-
-.user-console-push-status-trigger:hover,
-.user-console-push-status-trigger:focus-visible {
-  background: hsl(var(--warning) / 0.12);
-  color: hsl(var(--warning) / 0.96);
-  outline: none;
-}
-
-.user-console-push-status-trigger:focus-visible {
-  box-shadow: 0 0 0 2px hsl(var(--warning) / 0.22);
-}
-
-.user-console-push-status-spacer {
-  display: block;
-  width: 1.75rem;
-  height: 1.75rem;
-}
-
-.user-console-log-meta {
-  color: hsl(var(--muted-foreground));
-  font-size: 0.8rem;
-  line-height: 1.2;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  max-width: 100%;
-}
-
-.user-console-log-transport {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 10px;
-  width: 100%;
-}
-
-.user-console-log-transport-item {
-  display: inline-flex;
-  align-items: baseline;
-  gap: 4px;
-}
-
-.user-console-log-transport-item em {
-  font-style: normal;
-  color: hsl(var(--muted-foreground));
-  font-size: 0.72rem;
-  letter-spacing: 0.01em;
-}
-
-.user-console-log-transport-item strong {
-  font-family: 'JetBrains Mono', 'SFMono-Regular', Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace;
-}
-
-.user-console-log-result-line {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  min-width: 0;
-}
-
-.user-console-log-error {
-  color: hsl(var(--muted-foreground));
-  font-size: 0.8rem;
-  line-height: 1.2;
-  max-width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  min-width: 0;
-  flex: 1 1 auto;
-}
-
-.user-console-log-status.status-badge {
-  align-self: flex-start;
-  width: auto;
-  min-width: 0;
-}
-
 .jobs-table:not(.tokens-table):not(.admin-users-table) th:nth-child(7),
 .jobs-table:not(.tokens-table):not(.admin-users-table) td.jobs-message-cell {
   width: 24%;
