+++ web/src/styles/dark.css
@@ -89,90 +89,6 @@
   text-overflow: clip;
 }
 
-.user-console-log-card {
-  gap: 10px;
-}
-
-.user-console-log-card-head {
-  display: grid;
-  grid-template-columns: minmax(0, 1fr) auto;
-  align-items: start;
-  gap: 10px;
-}
-
-.user-console-log-card-request {
-  min-width: 0;
-  display: grid;
-  gap: 2px;
-}
-
-.user-console-log-card-request strong {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  font-weight: 700;
-}
-
-.user-console-log-card-request span {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  color: hsl(var(--muted-foreground));
-  font-size: 0.82rem;
-}
-
-.user-console-log-card-meta {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 6px 12px;
-  color: hsl(var(--muted-foreground));
-  font-family: 'JetBrains Mono', 'SFMono-Regular', Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace;
-  font-size: 0.82rem;
-}
-
-.user-console-log-card-meta time {
-  color: hsl(var(--foreground) / 0.82);
-}
-
-.user-console-log-card-error {
-  margin: 0;
-  color: hsl(var(--muted-foreground));
-  font-size: 0.86rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.user-console-mobile-status {
-  margin-left: auto;
-}
-
-.user-console-mobile-actions {
-  justify-content: flex-end;
-  gap: 8px;
-  flex-wrap: wrap;
-}
-
-.access-stats {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 12px;
-}
-
-.access-stats + .access-stats {
-  margin-top: 12px;
-}
-
-.access-stat {
-  border: 1px solid hsl(var(--border) / 0.82);
-  border-radius: 14px;
-  background: hsl(var(--card) / 0.92);
-  padding: 14px 18px;
-  min-width: 160px;
-  flex: 1 1 160px;
-}
-
 .access-stat h4,
 .access-stat-title {
   margin: 0;
